back to int, cause it works, and isn't uch a hassle to reconfigure
back to int

+++ pong-master/Pong/Pong/game.cpp
@@ -82,7 +82,7 @@ void game::Render()
 
 void game::RenderBall() 
 {
-	Ball = { (int)ballX, (int)ballY, 10, 10 };
+	Ball = { ballX, ballY, 10, 10 };
 	SDL_SetRenderDrawColor(renderer, 255, 255, 255, 0);
 	SDL_RenderFillRect(renderer, &Ball);
 }
